requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aioredis==1.3.1           # via channels-redis
amqp==5.0.3               # via kombu
apipkg==1.5               # via execnet
asgiref==3.3.1            # via channels, channels-redis, daphne, django
async-timeout==3.0.1      # via aioredis
attrs==20.3.0             # via automat, pytest, service-identity, twisted
autobahn==20.12.3         # via daphne
automat==20.2.0           # via twisted
autopep8==1.5.4           # via django-silk
backcall==0.2.0           # via ipython
billiard==3.6.3.0         # via celery
boto3==1.16.60            # via -r requirements.in
botocore==1.19.60         # via boto3, s3transfer
celery==5.0.5             # via -r requirements.in, django-celery-email
certifi==2020.12.5        # via requests, sentry-sdk
cffi==1.14.4              # via cryptography
channels-redis==3.2.0     # via -r requirements.in
channels==3.0.3           # via -r requirements.in, channels-redis
chardet==4.0.0            # via requests
click-didyoumean==0.0.3   # via celery
click-plugins==1.1.1      # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-plugins, click-repl, uvicorn
constantly==15.1.0        # via twisted
coverage==5.4             # via pytest-cov
cryptography==3.3.1       # via -r requirements.in, autobahn, pyopenssl, service-identity
daphne==3.0.1             # via channels
decorator==4.4.2          # via ipython
dj-database-url==0.5.0    # via -r requirements.in
django-anymail[mailgun]==8.1  # via -r requirements.in
django-appconf==1.0.4     # via django-celery-email
django-celery-email==3.0.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-extensions==3.1.0  # via -r requirements.in
django-model-utils==4.1.1  # via -r requirements.in
django-redis==4.12.1      # via -r requirements.in
django-silk==4.1.0        # via -r requirements.in
django-storages==1.11.1   # via -r requirements.in
django-taggit==1.3.0      # via -r requirements.in
django-turbo-response==0.0.28  # via -r requirements.in
django-widget-tweaks==1.4.8  # via -r requirements.in
django==3.1.5             # via -r requirements.in, channels, django-anymail, django-appconf, django-celery-email, django-model-utils, django-redis, django-silk, django-storages, django-taggit
execnet==1.8.0            # via pytest-xdist
factory-boy==3.2.0        # via -r requirements.in
faker==5.8.0              # via factory-boy
future==0.18.2            # via -r requirements.in
gprof2dot==2019.11.30     # via django-silk
gunicorn==20.0.4          # via -r requirements.in
h11==0.12.0               # via uvicorn
hiredis==1.1.0            # via aioredis
hyperlink==21.0.0         # via autobahn, twisted
idna==2.10                # via hyperlink, requests, twisted
incremental==17.5.0       # via twisted
iniconfig==1.1.1          # via pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements.in
jedi==0.18.0              # via ipython
jinja2==2.11.2            # via django-silk
jmespath==0.10.0          # via boto3, botocore
kombu==5.0.2              # via celery
markdown==3.3.3           # via -r requirements.in
markupsafe==1.1.1         # via jinja2
msgpack==1.0.2            # via channels-redis
packaging==20.8           # via pytest
parso==0.8.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.1.0             # via -r requirements.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.14    # via click-repl, ipython
psycopg2-binary==2.8.6    # via -r requirements.in
ptyprocess==0.7.0         # via pexpect
py==1.10.0                # via pytest, pytest-forked
pyasn1-modules==0.2.8     # via service-identity
pyasn1==0.4.8             # via pyasn1-modules, service-identity
pycodestyle==2.6.0        # via autopep8
pycparser==2.20           # via cffi
pygments==2.7.4           # via django-silk, ipython
pyhamcrest==2.0.2         # via twisted
pyopenssl==20.0.1         # via -r requirements.in, twisted
pyparsing==2.4.7          # via packaging
pytest-cov==2.11.1        # via -r requirements.in
pytest-django==4.1.0      # via -r requirements.in
pytest-forked==1.3.0      # via pytest-xdist
pytest-mock==3.5.1        # via -r requirements.in
pytest-randomly==3.5.0    # via -r requirements.in
pytest-xdist==2.2.0       # via -r requirements.in
pytest==6.2.2             # via -r requirements.in, pytest-cov, pytest-django, pytest-forked, pytest-mock, pytest-randomly, pytest-xdist
python-dateutil==2.8.1    # via botocore, django-silk, faker
pytz==2020.5              # via -r requirements.in, celery, django, django-silk
redis==3.5.3              # via -r requirements.in, django-redis
requests==2.25.1          # via django-anymail, django-silk
s3transfer==0.3.4         # via boto3
sentry-sdk==0.19.5        # via -r requirements.in
service-identity==18.1.0  # via twisted
six==1.15.0               # via automat, click-repl, cryptography, pyopenssl, python-dateutil
sorl-thumbnail==12.7.0    # via -r requirements.in
sqlparse==0.4.1           # via django, django-silk
text-unidecode==1.3       # via faker
toml==0.10.2              # via autopep8, pytest
traitlets==5.0.5          # via ipython
twisted[tls]==20.3.0      # via daphne
txaio==20.12.1            # via autobahn
unidecode==1.1.2          # via -r requirements.in
urllib3==1.26.3           # via botocore, requests, sentry-sdk
uvicorn==0.13.3           # via -r requirements.in
uvloop==0.14.0            # via -r requirements.in
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit
websockets==8.1           # via -r requirements.in
werkzeug==1.0.1           # via -r requirements.in
zope.interface==5.2.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
